Submition/requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.6
kiwisolver==1.4.8
matplotlib==3.10.0
mysql-connector-python==9.2.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
tzdata==2025.1
